+++ src/py/bbctrl/Ctrl.py
@@ -99,8 +99,10 @@ class Ctrl(object):
 
 
     def configure(self):
+        # Called from Comm.py after AVR vars are loaded
         # Indirectly configures state via calls to config() and the AVR
         self.config.reload()
+        self.state.init()
 
 
     def ready(self):